fixing mask projection

+++ util/utilIO.cpp
@@ -103,7 +103,7 @@ void ImgIO::getPtsFromMask(const cv::Mat &mask, std::vector<cv::Point2f> &pts_ve
 
   for(int r = 0; r < rows; r++){
     for(int c = 0; c < cols; c++){      
-      std::cout<< mask.at<float>(r,c) << std::endl;
+      std::cout<< mask.at<uchar>(r,c) << std::endl;
       if(mask.at<double>(r,c)>0){
 	pts_vector.push_back(cv::Point2f(r,c));
       }
